requirements.txt
python-dotenv
pandas
pymongo
numpy
certifi
pymongo
scikit-learn
pyyaml
dill
-e .